requirements.txt
torch==2.7.1
torchvision==0.22.1
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.3.0
numpy==1.26.4
pydantic==2.7.1
scikit-learn==1.5.0
python-multipart==0.0.9
loguru==0.7.2
transformers==4.41.2